+++ src/main/java/com/unirutas/auth/handlers/implementation/BasicAuthenticationHandler.java
@@ -1,18 +1,18 @@
-package com.unirutas.auth.implementation;
+package com.unirutas.auth.handlers.implementation;
 
-import com.unirutas.auth.handlers.AuthenticationHandler;
+import com.unirutas.auth.handlers.interfaces.IAuthenticationHandler;
 import com.unirutas.models.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
 /**
- * The SecurityPhraseAuthenticationHandler is responsible for user authentication using a security phrase.
- * It checks username, password, and security phrase for authentication.
+ * The BasicAuthenticationHandler is responsible for basic user authentication.
+ * It checks username and password for authentication.
  */
-public class SecurityPhraseAuthenticationHandler implements AuthenticationHandler {
-    private AuthenticationHandler successor;
-    private static final Logger logger = LoggerFactory.getLogger(SecurityPhraseAuthenticationHandler.class);
+public class BasicAuthenticationHandler implements IAuthenticationHandler {
+    private IAuthenticationHandler successor;
+    private static final Logger logger = LoggerFactory.getLogger(BasicAuthenticationHandler.class);
 
     /**
      * Authenticates the user based on the provided credentials and additional factors.
@@ -21,15 +21,14 @@ public class SecurityPhraseAuthenticationHandler implements AuthenticationHandle
      * @param username       The username for authentication.
      * @param password       The password for authentication.
      * @param phone          The phone number for phone or multi-factor authentication. (In this case is null)
-     * @param securityPhrase The security phrase for security phrase or multi-factor authentication.
+     * @param securityPhrase The security phrase for security phrase or multi-factor authentication. (In this case is null)
      * @return True if authentication is successful, false otherwise.
      */
     @Override
     public boolean authenticate(User user, String username, String password, String phone, String securityPhrase) {
-        if (user.getSecurityPhrase() != null && user.getPhone() == null){
-            logger.info("Validating credentials and security phrase...");
-            if (user.getUsername().equals(username) && user.getPassword().equals(password) && user.getSecurityPhrase().equals(securityPhrase)){
-                String message = "Successful authentication for "+ user.getName()+".";
+        if (user.getSecurityPhrase() == null && user.getPhone() == null){
+            logger.info("Validating credentials...");
+            if (user.getUsername().equals(username) && user.getPassword().equals(password)){
                 logger.info("Successful authentication for "+ user.getName()+".");
                 return true;
             } else {
@@ -42,12 +41,12 @@ public class SecurityPhraseAuthenticationHandler implements AuthenticationHandle
     }
 
     @Override
-    public AuthenticationHandler getSuccessor() {
+    public IAuthenticationHandler getSuccessor() {
         return successor;
     }
 
     @Override
-    public void setSuccessor(AuthenticationHandler successor) {
+    public void setSuccessor(IAuthenticationHandler successor) {
         this.successor = successor;
     }
 }
