+++ src/main/java/com/unirutas/auth/handlers/implementation/PhoneAuthenticationHandler.java
@@ -1,18 +1,18 @@
-package com.unirutas.auth.implementation;
+package com.unirutas.auth.handlers.implementation;
 
-import com.unirutas.auth.handlers.AuthenticationHandler;
+import com.unirutas.auth.handlers.interfaces.IAuthenticationHandler;
 import com.unirutas.models.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
 /**
- * The BasicAuthenticationHandler is responsible for basic user authentication.
- * It checks username and password for authentication.
+ * The PhoneAuthenticationHandler is responsible for user authentication using a phone.
+ * It checks username, password, and phone for authentication.
  */
-public class BasicAuthenticationHandler implements AuthenticationHandler {
-    private AuthenticationHandler successor;
-    private static final Logger logger = LoggerFactory.getLogger(BasicAuthenticationHandler.class);
+public class PhoneAuthenticationHandler implements IAuthenticationHandler {
+    private IAuthenticationHandler successor;
+    private static final Logger logger = LoggerFactory.getLogger(PhoneAuthenticationHandler.class);
 
     /**
      * Authenticates the user based on the provided credentials and additional factors.
@@ -20,15 +20,15 @@ public class BasicAuthenticationHandler implements AuthenticationHandler {
      * @param user           The user to authenticate.
      * @param username       The username for authentication.
      * @param password       The password for authentication.
-     * @param phone          The phone number for phone or multi-factor authentication. (In this case is null)
+     * @param phone          The phone number for phone or multi-factor authentication.
      * @param securityPhrase The security phrase for security phrase or multi-factor authentication. (In this case is null)
      * @return True if authentication is successful, false otherwise.
      */
     @Override
     public boolean authenticate(User user, String username, String password, String phone, String securityPhrase) {
-        if (user.getSecurityPhrase() == null && user.getPhone() == null){
-            logger.info("Validating credentials...");
-            if (user.getUsername().equals(username) && user.getPassword().equals(password)){
+        if (user.getPhone() != null && user.getSecurityPhrase() == null){
+            logger.info("Validating credentials and phone...");
+            if (user.getUsername().equals(username) && user.getPassword().equals(password) && user.getPhone().equals(phone)){
                 logger.info("Successful authentication for "+ user.getName()+".");
                 return true;
             } else {
@@ -41,12 +41,12 @@ public class BasicAuthenticationHandler implements AuthenticationHandler {
     }
 
     @Override
-    public AuthenticationHandler getSuccessor() {
+    public IAuthenticationHandler getSuccessor() {
         return successor;
     }
 
     @Override
-    public void setSuccessor(AuthenticationHandler successor) {
+    public void setSuccessor(IAuthenticationHandler successor) {
         this.successor = successor;
     }
 }
