+++ src/main/java/com/unirutas/auth/handlers/implementation/SecurityPhraseAuthenticationHandler.java
@@ -1,18 +1,18 @@
-package com.unirutas.auth.implementation;
+package com.unirutas.auth.handlers.implementation;
 
-import com.unirutas.auth.handlers.AuthenticationHandler;
+import com.unirutas.auth.handlers.interfaces.IAuthenticationHandler;
 import com.unirutas.models.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
 /**
- * The MultiFactorAuthenticationHandler is responsible for user authentication using a security phrase and phone.
- * It checks username, password, security phrase, and phone for authentication.
+ * The SecurityPhraseAuthenticationHandler is responsible for user authentication using a security phrase.
+ * It checks username, password, and security phrase for authentication.
  */
-public class MultiFactorAuthenticationHandler implements AuthenticationHandler {
-    private AuthenticationHandler successor;
-    private static final Logger logger = LoggerFactory.getLogger(MultiFactorAuthenticationHandler.class);
+public class SecurityPhraseAuthenticationHandler implements IAuthenticationHandler {
+    private IAuthenticationHandler successor;
+    private static final Logger logger = LoggerFactory.getLogger(SecurityPhraseAuthenticationHandler.class);
 
     /**
      * Authenticates the user based on the provided credentials and additional factors.
@@ -20,15 +20,16 @@ public class MultiFactorAuthenticationHandler implements AuthenticationHandler {
      * @param user           The user to authenticate.
      * @param username       The username for authentication.
      * @param password       The password for authentication.
-     * @param phone          The phone number for phone or multi-factor authentication.
+     * @param phone          The phone number for phone or multi-factor authentication. (In this case is null)
      * @param securityPhrase The security phrase for security phrase or multi-factor authentication.
      * @return True if authentication is successful, false otherwise.
      */
     @Override
     public boolean authenticate(User user, String username, String password, String phone, String securityPhrase) {
-        if (user.getPhone() != null && user.getSecurityPhrase() != null){
-            logger.info("Validating credentials, phone and security phrase...");
-            if (user.getUsername().equals(username) && user.getPassword().equals(password) && user.getPhone().equals(phone) && user.getSecurityPhrase().equals(securityPhrase)){
+        if (user.getSecurityPhrase() != null && user.getPhone() == null){
+            logger.info("Validating credentials and security phrase...");
+            if (user.getUsername().equals(username) && user.getPassword().equals(password) && user.getSecurityPhrase().equals(securityPhrase)){
+                String message = "Successful authentication for "+ user.getName()+".";
                 logger.info("Successful authentication for "+ user.getName()+".");
                 return true;
             } else {
@@ -41,12 +42,12 @@ public class MultiFactorAuthenticationHandler implements AuthenticationHandler {
     }
 
     @Override
-    public AuthenticationHandler getSuccessor() {
+    public IAuthenticationHandler getSuccessor() {
         return successor;
     }
 
     @Override
-    public void setSuccessor(AuthenticationHandler successor) {
+    public void setSuccessor(IAuthenticationHandler successor) {
         this.successor = successor;
     }
 }
